+++ shared/api-loaders.js
@@ -17,6 +17,30 @@ async function getProject(api, id, idType = 'id') {
   return { ...project, ...data };
 }
 
+const parseTeam = (team) => {
+  const ADMIN_ACCESS_LEVEL = 30;
+  const adminIds = team.teamPermissions.filter((user) => user.accessLevel === ADMIN_ACCESS_LEVEL);
+  team.adminIds = adminIds.map((user) => user.userId);
+  return team;
+};
+
+async function getTeam(api, id, idType = 'id') {
+  const team = await getSingleItem(api, `v1/teams/by/${idType}?${idType}=${encodeURIComponent(id)}`, id);
+  if (!team) return team;
+  const [users, pinnedProjects, projects, collections] = await Promise.all([
+    // load all users, need to handle pagination
+    getAllPages(api, `v1/teams/by/id/users?id=${team.id}&orderKey=createdAt&orderDirection=ASC&limit=100`),
+    getAllPages(api, `v1/teams/by/id/pinnedProjects?id=${team.id}&orderKey=createdAt&orderDirection=DESC&limit=100`),
+    getAllPages(api, `v1/teams/by/id/projects?id=${team.id}&orderKey=createdAt&orderDirection=DESC&limit=100`),
+    getAllPages(api, `v1/teams/by/id/collections?id=${team.id}&orderKey=createdAt&orderDirection=DESC&limit=100`),
+  ]);
+  team.users = users.sort((a, b) => new Date(a.teamPermission.updatedAt) - new Date(b.teamPermission.updatedAt));
+  team.projects = projects.sort((a, b) => new Date(b.updatedAt) - new Date(a.updatedAt));
+  team.teamPins = pinnedProjects.map((project) => ({ projectId: project.id }));
+  team.collections = collections;
+  return team && parseTeam(team);
+};
+
 async function getUser(api, id, idType = 'id') {
   const user = await getSingleItem(api, `v1/users/by/${idType}?${idType}=${encodeURIComponent(id)}`, id);
   if (!user) return user;
@@ -32,5 +56,6 @@ async function getUser(api, id, idType = 'id') {
 module.exports = {
   getCollection,
   getProject,
+  getTeam,
   getUser,
 };
